requirements.txt
streamlit
opencv-python
numpy
pandas
scikit-learn
matplotlib
seaborn
joblib
Pillow
